changing the type column of the ProfileField model to include a namespace `ProfileFieldTypes::`.
Typed profile_fields are now created like this:
    ProfileFieldTypes::Email.create( label: ..., value: ... )
    ProfileField.create( type: 'ProfileFieldTypes::Email', label: ..., value: ... )
    user.profile_fields.create( type: 'ProfileFieldTypes::Email', label: ..., value: ... )

The existant database entries have to be migrated using `rake db:migrate`.

Warning! After a while, the intermediate classes will be deleted. An automatic migration is then impossible.
But this shouldn't worry you, because this is only relevant for core developers in an early
dev stage.