Fixing remove button issue:
Clicking the remove button of a profile field has sent TWO delete requests instead of one.
This is because one has been sent by using jquery ujs (link_to remote:true) in
  app/views/shared/_add_profile_field_button.html.haml   and
  app/helpers/tool_buttons_helper.rb .
And the same thing was done parallely with javascript in
  app/assets/javascripts/profile_fields.js.coffee .

I've commented out the js one. Not sure which one should be used.