beginning to refactor validity range of dag_links.
We do not use acts_as_paranoid_dag anymore, since there are issues regarding the count attribute that may cause the database to break, which we have referred to as 'dag link error'.
Instead we first try to implement a validity range using scopes, which will have nothing to do with the destroy method.