travis: first run db:test:prepare in order to make sure the db is empty after running the migrations.
Actually, at the moment, the migrations appear to create the corporations_parent group.
If the database is filled initially, the database cleaner won't clean it, which is expected.
We should make sure the migrations do not create groups. The migrations are only allowed to modify data when changing the schema.