Changing the location of the file_store cache to tmp/app_cache. (Before: app/cache.)
We do this, because the assets cache, by default, is also stored in tmp/cache and is reset on each deploy by rake assets:precompile. But, we do not want to delete the application cache on each deployment.

See also this previous commit:
https://github.com/fiedl/wingolfsplattform/commit/224b8c1ffcb01a8c452f3166e915903b796107c9

Trello: https://trello.com/c/gqhbW3qq/646-performance-optimierung-von-gruppen