increasing devise timeout.
We are optimizing for users that use that application on their own computers. If they use it outside, they should use the 'Logout' button to quit their sessions. Otherwise, they should stay logged in if they do not log out. Therefore, I'm increasing the timeout of inactivity after that they are asked to re-login to a large value of 30 days.

Trello: https://trello.com/c/WbJHgb3b/262-devise-teil-2