fixing upload issues concerning pages with blog posts.
if there are blog posts on a page, the upload via drag and drop does not work
properly since it is not clear where to put the attachment.

therefore, motivate the user to visit the blog post directly and then upload
the attachment.

uservoice: https://wingolf.uservoice.com/admin/tickets/244