group member list export: making specs sensitive.
The specs have not been checking all list export types due to a syntax
mistake.

Always use `[‚foo‘, ‚bar‘].each { |str| … }` and do not use `for str in
[‚foo‘, ‚bar‘]`!

As a result, the specs have not been sensitive to an issue where the
birthday list could not be created due to a missing date of birth.

The spec is now fixed.