best_in_place: circumventing display_with issue.
The best_in_place gem, which we use for in-place editing, allows
formatters ("display methods")
that are applied to the values of in-place editing field. For example,
we use a markdown formatter
for certain textareas.

Unfortunately, the formatter definitions are stored in a class variable
by the gem, which is not
shared between the unicorn worker instances of the application.
Therefore, it is not ensured
that all display methods required are being loaded by the worker that
responds to the ajax request
of `respond_with_bip` in the controller.

An initializer circumvents this problem by defining all needed
definitions by hand in order to
ensure that all worker instances have the same definitions.

See also this issue on github:
https://github.com/bernat/best_in_place/issues/321

In order to circumvent this issue, we had to patch two methods of
best_in_place:

* `BestInPlace::BestInPlaceHelpers.best_in_place`
* `BestInPlace::ControllerExtensions.respond_bip_ok`

Trello:
https://trello.com/c/K4Uyw5QN/702-best-in-place-204-no-content-mit-unico
rn