events: fixing datetime issue.
We experience problems with the delocalize gem after upgrading our
rails version.
https://github.com/clemens/delocalize/issues/74

In order to avoid these issues, we trigger the parsing manually, for
the moment.
`Delocalize::LocalizedDateTimeParser.parse(string, Time).to_time`

In order to display the time in the timezone of the user, we set the
time zone in the
application controller as shown in railscast 106:
http://railscasts.com/episodes/106-time-zones-revised

This also converts the time back to UTC when posting the form.

Trello:
https://trello.com/c/5hvIZRfj/817-fix-delocalize-datums-und-uhrzeitangab
en