merit: adding calendar-uplink badge
trello: https://trello.com/c/Sl2cZspV/886-gamification

In this merit example, the badge is not granted from
`Merit::BadgeRules` but from the `EventsController#index`, since we
need `request.format` and `request.protocol` to determine whether the
badge should be granted.

Furthermore, this adds a short spec for earning the badge.