connected groups: implementing MembershipCollection#first_per_group
I thought, I would use this for group_members#index. But it turns out,
this dos not result in the expected behaviour.

In a super group, rather than seeing the date of joining the current
direct group, I want to see when the user joined one of the direct
groups the first time.