connected groups: Implementing MembershipCollection#join_validity_ranges_of_indirect_memberships.
This joins indirect memberships when they are of the same user and the
same group but different validity ranges:

    |-----------|                   first indirect membership in @group1
                |---------          second indirect membership in
@group2
    |---------------------          joined indirect membership

This is needed, since in `group_members#index` of indirect groups, the
users should be listed only once. They should be listed with the
earliest valid_from of the direct memberships. It’s not enough to use
`MembershipCollection#first_per_group`, because this would only
consider current direct memberships.