experimenting with api to fetch partials
While working on a view that helps admins to manage member data more
efficiently, we need to replace parts of the profile page
asynchronously.

One possibility would be to load some partials via an api.

It is unlikely that we stick to that approach at the moment. But this
commit documents some experiments.

Trello: https://trello.com/c/Rsoo0cFQ/968-deissner-liste-2015-bai3aes0