making galleria work with teaser boxes (2016-reiGha5u)
Clicking on the image of teaser boxes redirects to the teased page.

Adding a magnification-glass icon in the corner of the teaser image.
When the user clicks on the magnification icon, the gallery will open.

Do not show image thumbnails. If the first image is interesting enough
to see it large, the user can also see the other images. Otherwise, the
user does not need to know that there are other images as well.

2016-reiGha5u

Trello: https://trello.com/c/u5Fl2VmC/991