making you_tube work with galleria (2016-reiGha5u)
In teaser boxes, using the auto_html helper, the video url would just
be replaced by an iframe. The video would be rather small then.

Instead, use galleria to display the video. If the user clicks on the
video, a lightbox opens and the video begins to play.

2016-reiGha5u

Trello: https://trello.com/c/u5Fl2VmC/991