working on iweb layout (2016-top6zi8O)
Hide the heading of the first box, since it does not really fit in
there.

Image reflection in galleria.

Working on layout switch and layout setting in the home_page_settings.

Working on the attachments drop field for pages, which should only be
visible when the page is in edit mode, since the author knows that he
can upload something. In events, we keep the drop field outside to
encourage people to upload photos.

2016-top6zi8O