working on responsive layout in non-public pages (2016-top6zi8O)
Improve page tool buttons. Also, show a settings tool button for teaser
boxes.

Move the vertical nav into the same row as the box-configured cols in
order to have the stored col classes behave correctly.

Recognize the layout of the root page as default layout for non-navable
objects.

Show the vertical nav if it is to be shown for the parent to avoid that
the vertical nav „vanishes“ unexpectedly when visiting a child.

Hide other stuff in the vertical nav column on public pages, e.g. page
recommendations.

2016-top6zi8O

Trello: https://trello.com/c/u5Fl2VmC/991