officer boxes on pages (2016-poo8AePh)
Officer boxes can be shown on any Page. This is configurable in the
page settings.

Either show the page officers, e.g. people responsible for a certain
sub page. Or, show the officers of a certain group instead, e.g. the
officers of the organization represented on the page.

2016-poo8AePh

Trello: https://trello.com/c/u5Fl2VmC/991