replacing gmaps4rails by native google maps api (2016-reiGha5u)
Why?

(1) We had lots of javascript errors.
(2) By doing it manually we have a lot less unused code.
(3) It’s unclear which is the current version of gmaps4rails, or, if it
is maintained anymore.
(4) We can drop the fork of gmaps4rails we needed temporarily.
(5) We could resolve the issues concerning the bootstrap nav-tabs view.

This has become necessary with the homepage feature due to javascript
errors.

2016-reiGha5u

Trello: https://trello.com/c/u5Fl2VmC/991

And about possible map replaements:
https://trello.com/c/PQoO7Ga5/1007