sessions: SessionsController
The SessionController handles all sessions, now. Use `sessions#…`
rather than `devise/sessions#…` in the routes, now.

This way, we can be sure that the methods in the SessionsController are
the ones that are called.

But beware, there are `prepend_before_action` declarations in
https://github.com/plataformatec/devise/blob/master/app/controllers/devi
se/sessions_controller.rb, which might cancel the action before it is
called.