best in place: remove "empty field" restriction
Previously, we needed to prevent best_in_place form submitting an empty
result, because users accidentally removed information.

We are not sure if the users actually removed the text from the input
fields or if edit_mode falsely sent the information. Best-in-place does
have a mechanism to send an update request only if the value has
actually changed.

Anyway, some manual tests suggest that the issue is gone by now. And we
need to be able to update fields with empty values in issues#index.

Therefore, remove the restriction. But this has to be monitored closely
by our testers.

Trello: https://trello.com/c/vVrTElX7/1037-wingolfsblatter-issues