Patch `ApplicationRecord#first`.
This way `Group.first` returns the first group regardless of type.
`Corporation.first` returns the first corporation, not the first group.

I’m not sure why this is an issue in the first place. Please recheck
after migrating to rails 5.

A sensitive spec is: spec/models/concerns/user_corporations_spec.rb:46.

This is the first patch utilizing the `ApplicationRecord` abstract
class, which will be introduced with rails 5 anyway.