dag link repair for deleting the links of Groups::Everyone
I’d like to remove the links from and to the special group
`Groups::Everyone` in order to speed up everything.

The group has several thousand links that are not needed but add
overhead to a lot of caching and writing operations.

Everyone is in this group using a `members` override anyway. And nobody
uses this group as navigational anchor as it has been intended in the
concept phase.

For the steps to take care of in production, see
https://github.com/fiedl/wingolfsplattform/pull/74.

This, form a maintenance perspective, adds a counter repair job to the
convenience method `DagLink.repair`.

https://trello.com/c/6dNTE3FL/1-renew-cache