connect events to groups by association rather than dag link
in order to improve performance.

We don’t need the overhead anymore. The last reason to have the events
within the graph structure connected to the groups has been that this
was needed to calculate breadcrumbs and the navigational structure.

But as this mechanism now uses `Navable#parent`, which can easily
overridden for the `Event` class, this is no longer needed.

This should improve performance significantly since a lot of indirect
links are not needed anymore to be written and updated. Events can be
created much faster now.

Previously, events could have several parent groups, i.e. could be
shared amongst groups. But this has not been used a single time. Thus,
we drop this to make the association even faster.

Also, this fixes lots of problems: Attendees and contact people, which
are still connected to the event by a local graph, are disconnected
from the global graph, and thus, are not mistakenly confused with
status groups anymore, allowing us to drop some filtering overhead.

This commit also splits up the Event code into some concerns.

https://trello.com/c/6dNTE3FL/1-renew-cache