make `Membership` an STI type of `DagLink`
There are a couple of things to watch out for:

- The acts-as-dag gem does not create indirect links when calling
`Membership.create`. Thus, we needed to override `Membership.create`.
The same for `#destroy`.
- Creating a membership through `DagLink.create` rather than
`Membership.create` will return a `DagLink` with type `“Membership“`,
not a `Membership`. When patching this behaviour, bad things happened!
Thus, we left this alone. When calling `Membership.create` a
`Membership` is returned as expected.

https://trello.com/c/D3onnCh1/4-refactoring-memberships