caching: increase the cache expiration to 1 year.
Set this incredibly large. This way, we find out if we’ve missed a
trigger that resets a cache. And we don’t have to `rake cache:all`
every week.

The new system is designed to renew caches when they need to be renewed.

Main apps or even models may override this by implementing
`ApplicationRecord#new_caches_expire_in`.

https://trello.com/c/6dNTE3FL/1-renew-cache-2017-2016-uqu0sahl