sidekiq: retry jobs on the `:retry` queue.
In order not to block the regular queue for problematic jobs.

Need to reconfigure the main app’s `Procfile` to launch a separate
worker. By default, all jobs are handled by the same worker.

Probably should set the concurrency of the `:retry` queue to 1 to
handle one job at a time. This way, more resources are available to
problematic jobs, which might be a good idea for timeouts et cetera.

https://trello.com/c/cQMbH6X4/8-re-configure-sidekiq