improve splitting up jobs when running RenewCacheJob as background job.
Previously, the splitting up only happened when calling `renew_cache_later` directly on a record, because `running_in_background_job` has been set for a record, not globally.

Not, this also works for nested calls like calling `DagLink#renew_cache_later`, which renews caches also for the `ancestor` and the `descendant`.

This should prevent jobs from hanging around for several hours and blocking the queue!

https://trello.com/c/6dNTE3FL/1-renew-cache-2017-2016-uqu0sahl
https://trello.com/c/cQMbH6X4/8-re-configure-sidekiq