changing how profile field values are stored to increase performance without further caching logic.
Previously, `ProfileField#value` was overridden. In order to increase
performance, I tried to cache `value`. But this did not work very well
due to the complicated relation between parent and child profile fields.

Now, `value` is not overridden. The composed value is accessed by
`composed_value`. Before saving the parent, the `composed_value` is
copied to `value`. This way, the plain `value` column should be up to
date after every save.