gems for rails 5: acts-as-dag, refile
acts-as-dag has not had an update for quite a while. The master on
github works with rails 5, the released gem does not. I’ve forked the
repo in case fixes are needed. There are a couple of open pull requests
regarding rails 5.

Refile has also some issues and open pull requests regarding rails 5.
Also the unstable refile-mini_magick is needed. To get it to work, I
needed to include a fork of some of one of the contributors as the pull
request is not merged.

Maybe, we need to drop refile. But we can’t drop acts-as-dag. Maybe we
need to maintain.

https://trello.com/c/oIddSLU0/1161-rails-5