removing patches and workarounds that are no longer needed in rails 5.
Mark issues that still exist for rechecking when migrating to the next
rails version.

Removing the `has_secure_token` gem as this is now part of rails 5.

Removing the `transaction_retry` gem as the test did not raise any
deadlock issues locally.

Also, fixing breadcrumb replacement when renaming the object through
the ui.

https://trello.com/c/oIddSLU0/1161-rails-5