ui to select existing groups.
Rather than entering the group id, this allows to enter some part of
the name or of the breadcrumb path of a group and shows a list of
groups to select from.

We can use this to add existing groups to nav tree, but also for other
features like the settings view of home pages.

https://trello.com/c/TukXF9pS/1175-ui-zur-auswahl-existierender-gruppen
https://trello.com/c/8KJS2P15/1174-vertikales-menü-existierende-sub-grup
pen-einhängen