transfer galleria to yarn, update and patch.
Our old version of galleria is not compatible to the updated version of
jquery we are using after the rails-5 update. Therefore, we need to
update galleria.

Rather than just downloading the new source file of galleria, use
`yarn` to install the `galleria` node package.

In order to make galleria compatible to the asset pipeline, the
galleria theme css needs to be compiled into the main css, not loaded
asynchronously. Previously, we have just patched the theme file inline.
To make updates easier, we patch the `Galleria.addTheme` method here
separately.

https://github.com/worseisbetter/galleria

https://trello.com/c/7ad1FV92/535-galleria

https://trello.com/c/oIddSLU0/1161-rails-5