Trying to fix huge performance issue after rails-5 update.
Turns out, this issue has been due to not skipping filters when downloading attachments.
This commit introduces a separate controller for attachment downlaods in order to
separate this funcionality.

Probably, this has been the cause:
https://github.com/fiedl/your_platform/commit/2081b8ee2ee768a1fa59aec0dede49fd6176ef66

Also, this commit introduces some fast-route authorizations for downloading
thumbnails and logos to avoid a penaltiy on authorization there where no
information is contained that should be protected.

https://trello.com/c/i7obt7JX/1194-störung-server-sehr-träge
https://trello.com/c/oIddSLU0/1161-rails-5