better approach to box resizing.
This allows for different layouts to define their own allowed column
classes.

    // iweb.coffee
    App.permitted_bootstrap_column_classes = [
      'col-sm-12', 'col-sm-8', 'col-sm-4'
    ]

The desired column class is determined by minimizing the width
deviation.

Also, the improves the feel when resizing a column, because the guess
with column class is desired is far more accurate compared to the
previous approach.

https://trello.com/c/UVpa6ee7/1216-inbetriebnahme-des-homepage-moduls-fü
r-wingolforg-2017-zu8ahm2i