neo4j: `Group#descendant_membership_ids` include the memberships of officer groups if the parent group is an officer group.
Why the distinction? When the parent group is an officer group, we are interested in the members of the officer group.

When, on the other hand, the parent group has officer groups as descendants, we don't want to include the members of the officer groups, because there could be officers thare are no regular members. They should not be considered as `descendant_members`.

Also fixing on the way: neo4j-variable-length relationships now support leaving out the upper limit

https://graphaware.com/graphaware/2015/05/19/neo4j-cypher-variable-length-relationships-by-example.html

https://trello.com/c/LFd4TZ1F/1375-ämterhistorie-bearbeiten-graph-gibt-keine-memberships-zurück