mailing lists: handle emails with several `X-Original-To` headers.
Consider an email sent to (`RCTP TO`) aktivitas@example.com. This address is configured to forward the message to mailgate@example.io. This address is configured to relay the message to our application.

Both addresses, aktivitas@example.com and mailgate@example.io are added as `X-Original-To` headers by the postfix instances. But only the first one is the one to compare to our mailing list register to find out the recipient group.

https://trello.com/c/zfdJNwBV/1385-mail-loop-in-e-mail-verteilern