remove coupling of `root#index` and the `Page.intranet_root`.
Previously, the intranet start page, handled by the `root#index` action, has been identical to the page with the flag `:intranet_root`.

Now, I'm decoupling this, in order to reuse the page without moving the child pages and groups.

We might reactivate the coupling later when the graph database is ready and it would make no difference to move the children to another node.

https://trello.com/c/GIrENd2O/64-seite-für-den-bund